+++ src/com/ylbms/base/single/service/SpectypeService.java
@@ -34,7 +34,7 @@ public class SpectypeService {
 		spectypeDao.save(spectype);
 	}
 
-	public SpectypeInfo getSpetypeById(int id) {
+	public SpectypeInfo getSpetypeById(long  id) {
 		return spectypeDao.get(id);
 	}
 
@@ -57,7 +57,7 @@ public class SpectypeService {
 	 * @param id
 	 */
 	@Transactional(readOnly = false)
-	public void deleteSpectype(int id) {
+	public void deleteSpectype(Long id) {
 		spectypeDao.delete(id);
 		// spectypeDao.findPage(page, filters)
 	}
