+++ src/com/ylbms/base/single/web/controller/SpectypeController.java
@@ -102,7 +102,7 @@ public class SpectypeController {
 
 	@RequestMapping(value = "/delete/{id}")
 	@ResponseBody
-	public Map<String, Object> delSpectype(@PathVariable("id") int id) {
+	public Map<String, Object> delSpectype(@PathVariable("id") Long id) {
 		try {
 			spectypeService.deleteSpectype(id);
 			return DwzUtil.dialogAjaxDone(DwzUtil.OK);
